[FIX] #111 seperated notification page component
