+++ src/components/notification/NotificationEmpty.vue
@@ -1,0 +1,5 @@
+<template>
+  <div class="flex justify-center items-center text-gray-400 py-16">
+    <img src="../../assets/images/logo.png" class="w-32 h-auto mt-28" />
+  </div>
+</template>
