+++ src/components/notification/NotificationLoading.vue
@@ -1,0 +1,3 @@
+<template>
+  <div class="text-center text-gray-500 py-4">불러오는 중…</div>
+</template>
